+++ scripts/Backstabs/settingsGlobal.lua
@@ -1,10 +1,10 @@
 local I = require('openmw.interfaces')
 
 I.Settings.registerGroup {
-    key = 'SettingsBackstabs_general',
+    key = 'SettingsBackstabs_toggles',
     page = 'Backstabs',
     l10n = 'Backstabs',
-    name = 'general_group_name',
+    name = 'toggles_group_name',
     permanentStorage = true,
     settings = {
         {
@@ -14,6 +14,55 @@ I.Settings.registerGroup {
             default = true,
         },
         {
+            key = 'enableMeleeWeapons',
+            name = 'enableMeleeWeapons_name',
+            renderer = 'checkbox',
+            default = true,
+        },
+        {
+            key = 'enableRangedWeapons',
+            name = 'enableRangedWeapons_name',
+            renderer = 'checkbox',
+            default = true,
+        },
+        {
+            key = 'enableH2h',
+            name = 'enableH2h_name',
+            renderer = 'checkbox',
+            default = true,
+        },
+        {
+            key = 'playerCanBeBackstabbed',
+            name = 'playerCanBeBackstabbed_name',
+            renderer = 'checkbox',
+            default = true,
+        },
+        {
+            key = 'enableSpecialWeaponInstakill',
+            name = 'enableSpecialWeaponInstakill_name',
+            description = 'enableSpecialWeaponInstakill_description',
+            renderer = 'checkbox',
+            default = false,
+        },
+        {
+            key = 'debugMode',
+            name = 'debugMode_name',
+            description = 'debugMode_description',
+            renderer = 'checkbox',
+            default = false,
+        },
+    }
+}
+
+I.Settings.registerGroup {
+    key = 'SettingsBackstabs_values',
+    page = 'Backstabs',
+    l10n = 'Backstabs',
+    name = 'values_group_name',
+    description = "values_group_description",
+    permanentStorage = true,
+    settings = {
+        {
             key = 'mode',
             name = 'mode_name',
             description = 'mode_description',
@@ -22,6 +71,7 @@ I.Settings.registerGroup {
                 l10n = "Backstabs",
                 items = {
                     "Mixed",
+                    "Gradual",
                     "Skill-only",
                     "Flat-only",
                     "Instakill"
@@ -34,14 +84,22 @@ I.Settings.registerGroup {
             name = 'flatMult_name',
             renderer = 'number',
             integer = false,
-            default = 5,
+            default = 1.5,
         },
         {
             key = 'sneakMult',
             name = 'sneakMult_name',
             renderer = 'number',
             integer = false,
-            default = 0.1,
+            default = 0.05,
+        },
+        {
+            key = 'gradualStep',
+            name = 'gradualStep_name',
+            renderer = 'number',
+            integer = true,
+            default = 25,
+            min = 1
         },
         {
             key = 'npcFov',
@@ -60,13 +118,6 @@ I.Settings.registerGroup {
             integer = false,
             default = 0.25,
             min = 0.01
-        },
-        {
-            key = 'debugMode',
-            name = 'debugMode_name',
-            description = 'debugMode_description',
-            renderer = 'checkbox',
-            default = false,
         },
     }
 }