+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Backstabs
 
+## 2.0
+
+### Features
+
+- H2h now can backtab too (somehow)
+- Player now can get backstabbed too
+- Combat checks now apply to players too
+- Added "Gradual" mode, akin to the later games
+- Added instakill backstabs for special weapons. Special weapons are:
+  - Mehrunes' Razor
+- Added separate toggles for backstabs with melee weapons, ranged weapons and h2h
+- Settings text tweaks
+
+### Fixes
+
+- Backstab detection now actually works
+- Toned down default mod values to be more reasonable
+
 ## 1.0
 
 Initial release.